Ex 1 Task 3 Split PowerShell code into two part
The copy and pasting of the PowerShell code was not working because the Read-Host did not stop for the user's input, instead the next lines were used as input and the resource group could not get created. For most users in the Cloud Shell that caused issues, on a local computer with normal copy and paste it seemed to work. Have split this to avoid an issue here.