Update Lab 7 instructions to delete ALL resource groups created
The student creates a resource group named az400m07l01-RG. The lab instructions correctly delete the resource group. 
But when students run the pipeline, an additional resource group is created named 'azurekeyvaultlab.' It contains the web app and a database.
This is not deleted in the instructions.  This commit mimics the original delete steps to remove the extra resource group.